# packages/shared/src/generated/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/cart/items": {
      "post": {
        "operationId": "OrdersController_addCartItem_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/cart/items/{menuItemId}": {
      "patch": {
        "operationId": "OrdersController_updateCartQty_v1",
        "parameters": [
          {
            "name": "menuItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartQtyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "delete": {
        "operationId": "OrdersController_removeCartItem_v1",
        "parameters": [
          {
            "name": "menuItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/cart/coupon": {
      "post": {
        "operationId": "OrdersController_applyCoupon_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyCouponDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "delete": {
        "operationId": "OrdersController_removeCoupon_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/cart/prescription": {
      "post": {
        "operationId": "OrdersController_uploadPrescription_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/checkout": {
      "post": {
        "operationId": "OrdersController_checkout_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/v1/orders/store/{orderId}/prescription/verify": {
      "post": {
        "operationId": "OrdersController_verifyPrescription_v1",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionVerificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/store/{orderId}/substitutions": {
      "post": {
        "operationId": "OrdersController_proposeSubstitutions_v1",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubstitutionProposalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/store/{orderId}/accept": {
      "post": {
        "operationId": "OrdersController_acceptStoreOrder_v1",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepTimeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/store/{orderId}/prepare": {
      "post": {
        "operationId": "OrdersController_prepareStoreOrder_v1",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepTimeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/v1/orders/{orderId}/substitutions": {
      "post": {
        "operationId": "OrdersController_decideSubstitutions_v1",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubstitutionDecisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{orderId}/cancel": {
      "post": {
        "operationId": "OrdersController_cancelOrder_v1",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{orderId}/rating": {
      "post": {
        "operationId": "OrdersController_rateOrder_v1",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/v1/orders/{orderId}": {
      "get": {
        "operationId": "OrdersController_getOrder_v1",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "MoveX API",
    "description": "Generated OpenAPI contract for MoveX clients",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-movex_session"
      }
    },
    "schemas": {
...
      "CartItemDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCartQtyDto": {
        "type": "object",
        "properties": {}
      },
      "ApplyCouponDto": {
        "type": "object",
        "properties": {}
      },
      "PrescriptionUploadDto": {
        "type": "object",
        "properties": {}
      },
      "CheckoutDto": {
        "type": "object",
        "properties": {}
      },
...
      "PrepTimeDto": {
        "type": "object",
        "properties": {}
      },
      "PrescriptionVerificationDto": {
        "type": "object",
        "properties": {}
      },
      "SubstitutionProposalDto": {
        "type": "object",
        "properties": {}
      },
      "SubstitutionDecisionDto": {
        "type": "object",
        "properties": {}
      },
...
      "CancelOrderDto": {
        "type": "object",
        "properties": {}
      },
      "RatingDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}